Refactor places models structure and add field for calculated area
- Split `Site` model into a new `sites.py` file for better modularity.
- Add `calculated_area_m2` field to the `Neighborhood` model to store
  calculated area in square meters.
- Update `__init__.py` to include imports for the new models structure.
- Rename `models.py` to `spatial.py` to better reflect its contents.
