+++ backend/apps/places/models/__init__.py
@@ -1,0 +1,12 @@
+from .sites import Site
+from .spatial import Country, Department, Municipality, Locality, Neighborhood
+
+
+__all__ = [
+    "Country",
+    "Department",
+    "Municipality",
+    "Locality",
+    "Neighborhood",
+    "Site",
+]
